third_party/blink/web_tests/fast/sub-pixel/computedstylemargin-expected.txt

PASS style.marginTop is '0.5px'
PASS style.marginRight is '1.5px'
PASS style.marginBottom is '2.5px'
PASS style.marginLeft is '3.5px'
PASS style.marginLeft is '0.5px'
PASS style.marginRight is '0.5px'
PASS successfullyParsed is true

TEST COMPLETE
